requirements.txt
azure-core==1.33.0
azure-storage-blob==12.25.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
pycparser==2.22
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
style==1.1.0
typing_extensions==4.13.2
update==0.0.1
urllib3==2.4.0
Werkzeug==3.1.3
